Region interface can now query cloud for flavors.
Support for the flavors/ endpoint added to the Region interface.
The code is structurally quite similar to the Images code.

As an aside:

I'm beginning to think that Region, et. al. should not be interfaces.
It's not very Go-like, and it seems like I should be doing something
different.  But, I'll defer future refactoring for a different patch.

Squashed commit of the following:

commit 491b1b215bd7ebc5c9411a84b5b426eb18c4d6d1
Author: Samuel A. Falvo II <sam.falvo@rackspace.com>
Date:   Fri Apr 26 17:21:26 2013 -0700

    Test refactoring

commit 3425c485de631e2ef4f63fbddf6677f5f35ac8f0
Author: Samuel A. Falvo II <sam.falvo@rackspace.com>
Date:   Fri Apr 26 16:55:33 2013 -0700

    godoc comments

commit bf8c06daf6a6faa1a0485412eac8fc0c60cdb0f9
Author: Samuel A. Falvo II <sam.falvo@rackspace.com>
Date:   Fri Apr 26 16:46:42 2013 -0700

    Source code formatting

commit 0722f4f14da425470082f4386759ecb8dafda8c4
Author: Samuel A. Falvo II <sam.falvo@rackspace.com>
Date:   Fri Apr 26 16:45:16 2013 -0700

    Code for listing flavors added.