Bug fixes and initial cloud server support
This commit refactors the RESTful accesses into a new project, called
perigee.  Perigee is intended to abstract all HTTP-level accesses.
Perigee is an open-by-default project.

Gorax itself now supports listing a set of servers.  Much of the
groundwork for creating servers now exists via the Perigee project as
well.

Documentation is improved as well.

Squashed commit of the following:

commit 35223e7d4ac5462484e7d542f11ab178460f3ef5
Author: Samuel A. Falvo II <sam.falvo@rackspace.com>
Date:   Thu May 2 16:37:40 2013 -0700

    Bug fixes and initial support for cloud servers.

commit 5b4024d1745fecbb952c08275862c06b58bfa135
Author: Samuel A. Falvo II <sam.falvo@rackspace.com>
Date:   Thu May 2 13:55:43 2013 -0700

    Updated to use new Get/Post in perigee.

    Many of the parameters provided are technically optional in the generic
    REST call sense, so they've been moved to an Options structure.

commit 7d8ae928c95f37c25df34258a14f9548f575c986
Author: Samuel A. Falvo II <sam.falvo@rackspace.com>
Date:   Thu May 2 13:16:58 2013 -0700

    Cleaning up redundant code