Added CreateServer to Region interface.
I did not add any tests for this API because it just re-uses
infrastructure which is already tested elsewhere.  The only differences
are the JSON request bodies and responses, which encoding/json handles
for us in a type-safe manner, and the specific endpoint and request
method, which Perigee handles for us.