Fix Servers() to give details about servers.
Previously, Servers() used the /servers endpoint.  This endpoint only
gives links to server details for each server the user has access to.
Altering the code to query /servers/detail _without_ specifying any
specific ID, we get a complete list of detailed server records.  This
prevents the API thrash one could expect when it comes to iterating over
a large number of servers.

N.B.: It would be nice if the documentation were formatted more clearly
on this subject.  The info was there, but for _weeks_, I just missed it
completely.  I only noticed the correct approach to this by chance when
studying listing 2.2 of
http://docs.rackspace.com/servers/api/v2/cs-devguide/content/List_Servers-d1e2078.html.